Move render functionality into module
